Calm luh update
I GOT A NEW PC OMGGGGGGG. words cannot contain my excitement on acquiring a new pc. And i built it all by my self. anyway about the game, i implemented some navigation to the enemy so that it can follow the player. Its tolerable for right now but  but it needs more work but ill handle that later. small update because of the ovious set back :)))))