Im not seeing enough movement
helo, today was ight i added some assets, fixed the fixed the car interactions with the items. I tried to fix the foot step audio but that was a bust. anyway gtk game jam is tomorrow so i really was just preparing for that so I'm hyped. anyways doing good.