Organize and Inv added and GAME JAM YEAAHHHHHHHHHH
Well I did it I did a my officially completed my first game jam it was rough but I did it and I'm so proud. once I was done I came back an organized some files and started to add a inventory, had to shift every thing but it works as intended. I plan to add a drag and drop system later for more interactivity. I also added a song that I made earlier this morning its more like a beat but i can always change it later. That all and doing well so far