Amongus in real life, sus,sus
Did a lot this time around, I changed some lighting, added a player hub, added a door, changed the car look, enemy look, player has a 2d hand and I also was going to add 3d enemies but since i aint got time for allat Im making the enemies 2d which in my opinion is way better anyway.  I added a weapon that has a bug with the damage but i found that is is caused by the for loop in the player script, in the manage animations function. The game is actually coming along now fellin good